Start splitting up grpal-cdk-stack
