+++ src/js/grapl-cdk/lib/grapl-cdk-stack.ts
@@ -15,7 +15,6 @@ import * as path from 'path';
 
 import { Service } from './service';
 import { UserAuthDb } from './userauthdb';
-import { HistoryDb } from './historydb';
 import { EventEmitter } from './event_emitters';
 import { RedisCluster } from './redis';
 import { EngagementNotebook } from './engagement';
@@ -27,189 +26,9 @@ import { OperationalAlarms, SecurityAlarms } from './alarms';
 import { Watchful, WatchedOperation } from 'cdk-watchful';
 import { SchemaDb } from './schemadb';
 import { PipelineDashboard } from './pipeline_dashboard';
-
-interface SysmonGraphGeneratorProps extends GraplServiceProps {
-    writesTo: s3.IBucket;
-}
-
-class SysmonGraphGenerator extends cdk.NestedStack {
-    readonly service: Service;
-
-    constructor(
-        parent: cdk.Construct,
-        id: string,
-        props: SysmonGraphGeneratorProps
-    ) {
-        super(parent, id);
-
-        const bucket_prefix = props.prefix.toLowerCase();
-        const sysmon_log = new EventEmitter(
-            this,
-            bucket_prefix + '-sysmon-log'
-        );
-
-        const event_cache = new RedisCluster(this, 'SysmonEventCache', props);
-        event_cache.connections.allowFromAnyIpv4(ec2.Port.allTcp());
-
-        this.service = new Service(this, id, {
-            prefix: props.prefix,
-            environment: {
-                BUCKET_PREFIX: bucket_prefix,
-                EVENT_CACHE_ADDR: event_cache.cluster.attrRedisEndpointAddress,
-                EVENT_CACHE_PORT: event_cache.cluster.attrRedisEndpointPort,
-            },
-            vpc: props.vpc,
-            reads_from: sysmon_log.bucket,
-            subscribes_to: sysmon_log.topic,
-            writes_to: props.writesTo,
-            version: props.version,
-            watchful: props.watchful,
-            metric_forwarder: props.metricForwarder,
-        });
-
-        this.service.event_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(parseInt(event_cache.cluster.attrRedisEndpointPort))
-        );
-
-        this.service.event_retry_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(parseInt(event_cache.cluster.attrRedisEndpointPort))
-        );
-    }
-}
-
-interface OSQueryGraphGeneratorProps extends GraplServiceProps {
-    writesTo: s3.IBucket;
-}
-
-class OSQueryGraphGenerator extends cdk.NestedStack {
-    constructor(
-        parent: cdk.Construct,
-        id: string,
-        props: OSQueryGraphGeneratorProps
-    ) {
-        super(parent, id);
-
-        const bucket_prefix = props.prefix.toLowerCase();
-        const osquery_log = new EventEmitter(
-            this,
-            bucket_prefix + '-osquery-log'
-        );
-
-        const event_cache = new RedisCluster(this, 'OSQueryEventCache', props);
-        event_cache.connections.allowFromAnyIpv4(ec2.Port.allTcp());
-
-        const service = new Service(this, id, {
-            prefix: props.prefix,
-            environment: {
-                BUCKET_PREFIX: bucket_prefix,
-                EVENT_CACHE_ADDR: event_cache.cluster.attrRedisEndpointAddress,
-                EVENT_CACHE_PORT: event_cache.cluster.attrRedisEndpointPort,
-            },
-            vpc: props.vpc,
-            reads_from: osquery_log.bucket,
-            subscribes_to: osquery_log.topic,
-            writes_to: props.writesTo,
-            version: props.version,
-            watchful: props.watchful,
-            metric_forwarder: props.metricForwarder,
-        });
-
-        service.event_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(parseInt(event_cache.cluster.attrRedisEndpointPort))
-        );
-
-        service.event_retry_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(parseInt(event_cache.cluster.attrRedisEndpointPort))
-        );
-    }
-}
-
-export interface NodeIdentifierProps extends GraplServiceProps {
-    writesTo: s3.IBucket;
-}
-
-class NodeIdentifier extends cdk.NestedStack {
-    readonly bucket: s3.Bucket;
-    readonly topic: sns.Topic;
-    readonly service: Service;
-
-    constructor(parent: cdk.Construct, id: string, props: NodeIdentifierProps) {
-        super(parent, id);
-
-        const history_db = new HistoryDb(this, 'HistoryDB', props);
-
-        const bucket_prefix = props.prefix.toLowerCase();
-        const unid_subgraphs = new EventEmitter(
-            this,
-            bucket_prefix + '-unid-subgraphs-generated'
-        );
-        this.bucket = unid_subgraphs.bucket;
-        this.topic = unid_subgraphs.topic;
-
-        const retry_identity_cache = new RedisCluster(
-            this,
-            'NodeIdentifierRetryCache',
-            props
-        );
-        retry_identity_cache.connections.allowFromAnyIpv4(ec2.Port.allTcp());
-
-        this.service = new Service(this, id, {
-            prefix: props.prefix,
-            environment: {
-                BUCKET_PREFIX: bucket_prefix,
-                RETRY_IDENTITY_CACHE_ADDR:
-                    retry_identity_cache.cluster.attrRedisEndpointAddress,
-                RETRY_IDENTITY_CACHE_PORT:
-                    retry_identity_cache.cluster.attrRedisEndpointPort,
-                STATIC_MAPPING_TABLE: history_db.static_mapping_table.tableName,
-                DYNAMIC_SESSION_TABLE:
-                    history_db.dynamic_session_table.tableName,
-                PROCESS_HISTORY_TABLE: history_db.proc_history.tableName,
-                FILE_HISTORY_TABLE: history_db.file_history.tableName,
-                INBOUND_CONNECTION_HISTORY_TABLE:
-                    history_db.inbound_connection_history.tableName,
-                OUTBOUND_CONNECTION_HISTORY_TABLE:
-                    history_db.outbound_connection_history.tableName,
-                NETWORK_CONNECTION_HISTORY_TABLE:
-                    history_db.network_connection_history.tableName,
-                IP_CONNECTION_HISTORY_TABLE:
-                    history_db.ip_connection_history.tableName,
-                ASSET_ID_MAPPINGS: history_db.asset_history.tableName,
-            },
-            vpc: props.vpc,
-            reads_from: unid_subgraphs.bucket,
-            subscribes_to: unid_subgraphs.topic,
-            writes_to: props.writesTo,
-            retry_code_name: 'node-identifier-retry-handler',
-            version: props.version,
-            watchful: props.watchful,
-            metric_forwarder: props.metricForwarder,
-        });
-
-        history_db.allowReadWrite(this.service);
-
-        this.service.event_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(
-                parseInt(retry_identity_cache.cluster.attrRedisEndpointPort)
-            )
-        );
-
-        this.service.event_retry_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(
-                parseInt(retry_identity_cache.cluster.attrRedisEndpointPort)
-            )
-        );
-
-        this.service.event_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(443),
-            'Allow outbound to S3'
-        );
-        this.service.event_retry_handler.connections.allowToAnyIpv4(
-            ec2.Port.tcp(443),
-            'Allow outbound to S3'
-        );
-    }
-}
+import { SysmonGraphGenerator } from './services/sysmon_graph_generator';
+import { OSQueryGraphGenerator } from './services/osquery_graph_generator';
+import { NodeIdentifier } from './services/node_identifier';
 
 export interface MetricForwarderProps extends GraplServiceProps {
     // nothing yet
